# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bondhon Matrimony API",
        "description": "Full-stack matrimony platform API. All protected endpoints require Bearer token from Sanctum.",
        "contact": {
            "email": "admin@bondhon.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get authenticated user's own profile",
                "operationId": "dffac35700b3186de8d19cc212fe0285",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update own profile (basic info + sub-details)",
                "operationId": "d682da2fd566e34f597e169ac582ad3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/{profileId}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "View another user's profile by profile ID (e.g. BON-000001)",
                "operationId": "dc690aac5aeaf3d6686cbb83ff5b9090",
                "parameters": [
                    {
                        "name": "profileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "BON-000001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile retrieved"
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "403": {
                        "description": "This profile is not visible to you"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/preferences": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update partner preferences",
                "operationId": "42abd543298abc6bfc7a53f94d2631f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/completion": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile completion status",
                "operationId": "9269cc8461c945a7997b7cfbe83da5d1",
                "responses": {
                    "200": {
                        "description": "Profile completion status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/photos": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Upload a profile photo",
                "operationId": "5b97d01c2bb1d700c0287e5cfbae26b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_private": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo uploaded, pending moderation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/photos/{photoId}": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete a profile photo",
                "operationId": "1a735bfb798f542528fc1855992e4301",
                "parameters": [
                    {
                        "name": "photoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo deleted"
                    },
                    "404": {
                        "description": "Photo not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/profile/photos/{photoId}/primary": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Set a photo as primary",
                "operationId": "2ac9110765add26e0e082fcd5e46801a",
                "parameters": [
                    {
                        "name": "photoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary photo set"
                    },
                    "403": {
                        "description": "Photo must be approved first"
                    },
                    "404": {
                        "description": "Photo not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}